requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.11
django-datetimepicker==3.14
django-markdownx==2.0.28
django-widget-tweaks==1.4.3
django-allauth==0.38.0
django-braces==1.13.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.9.1
html2text==2018.1.9
html2markdown==0.1.7
httplib2==0.12.0
idna==2.8
mistletoe==0.7.1
oauth2==1.9.0.post1
oauthlib==3.0.0
PyJWT==1.7.1
python-dateutil==2.7.5					  
python-decouple==3.1
python-social-auth==0.2.21
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==1.2.0
socialauth==0.2.0
soupsieve==1.8
urllib3>=1.24.2
virtualenv==16.2.0
Pillow==6.2.1
mysqlclient
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1